md2img/requirements_api.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
playwright>=1.40.0
pillow>=10.0.0
markdown>=3.5.0
pydantic>=2.0.0
requests>=2.31.0

